# pharma-platform/server/docs/openapi.yaml
openapi: 3.1.0
info:
  title: PharmaGuard API
  version: 1.0.0
  description: |
    Pharmaceutical Data Integrity & Compliance Platform.
    FDA 21 CFR Part 11 compliant electronic records, audit trails, 
    e-signatures, batch tracking, and document control APIs.
  contact:
    name: PharmaGuard Engineering
    email: api@pharmaguard.io
  license:
    name: Proprietary
servers:
  - url: http://localhost:3001/api/v1
    description: Local Development
  - url: https://api.pharmaguard.io/api/v1
    description: Production
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/login
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for watch agents and external integrations
...
paths:
...
  /documents:
    get:
      tags: [Documents]
      summary: List quality documents
      parameters:
        - { name: category, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string } }
        - { name: search, in: query, schema: { type: string } }
      responses:
        '200': { description: Paginated document list }

    post:
      tags: [Documents]
      summary: Create new document
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [documentNumber, title, category]
              properties:
                documentNumber: { type: string, example: "SOP-QC-003" }
                title: { type: string }
                category: { type: string }
                department: { type: string }
                tags: { type: array, items: { type: string } }
      responses:
        '201': { description: Document created }
  /documents/{id}:
    get:
      tags: [Documents]
      summary: Get document with versions and signatures
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Full document with version history }
  /documents/{id}/approve:
    patch:
      tags: [Documents]
      summary: Approve document with e-signature
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [intent]
              properties:
                intent: { type: string, example: "Reviewed and approved for use" }
                effectiveDate: { type: string, format: date }
      responses:
        '200': { description: Document approved with e-signature }

  # ─── Batches ───────────────────────────────────────────
...